HADOOP-19256. S3A: Support S3 Conditional Writes: rename builder()
Outline of a RenameOperation class which creates a builder which would
ultimately initiate a rename.

It's not *that* complex an operation API-wise.
The pain will be in implementing it consistently, especially if we make that
non-posix rename into empty dir feature optional with an withLegacyPathFixup()
flag.

HDFS would need to implement this server side.
Without that, it could be done for hdfs on the client.

ABFS and S3A could both benefit

Change-Id: I030842543dc60cb4cba3716a6f9d87dd986dffc0
ABFS: etag/file status for better file rename, saved HEAD
S3A: save on HEAD/LIST, progress callbacks (hello distCp!),