HADOOP-19363. Fixing audit tests
Most of this PR is trying to improve debugging of the auditor
invocation plane on the assumption that those flags being passed
down from the factory were causing problems.
None of those changes did any good, though they did marginally
improve debugging.

The actual problem was ordering of component startup during FS
init: the Auditor must be live before the AWS client is initialized.
Moved back to the right place and improved documentation.

Also: added a test to verify that setting flags would disable
the span checks, which is what we now require.

Change-Id: I108116f0775b71b1cf1c9a2bd5c95727f24f37bb