HADOOP-19303. VectorIO API
Adds a new method with a release method too

readVectored(List<? extends FileRange> ranges,
      IntFunction<ByteBuffer> allocate,
      Consumer<ByteBuffer> release)

This is return buffers to pools even in failures.

The default implementation hands back to readVectored/2,
so that existing custom implementations of that will get
invoked.

There is a weakness there: it means implementations which
don't implement readVectored/2 will not have buffer release
working.

Checksum, Buffer and RawLocal all do this.

S3A, ABFS and HDFS don't do this.

Change-Id: I26fa7214014482e39af6a9212d2e30dd0fc6d429