HADOOP-19546. Include cipher feature for HttpServer2 and SSLFactory
Currently, we have a feature to exclude weak ciphers from HttpServer2 and SSLFactory using the ssl.server.exclude.cipher.list property.
With this feature, we can also define an inclusion list of ciphers using the ssl.server.include.cipher.list property.
If the inclusion list is populated, any cipher not present in the list will not be allowed.
If a cipher is present in both the exclusion and inclusion lists, it will be excluded.
Note that SSLFactory does not support regex-based cipher patterns, unlike HttpServer2.