HADOOP-18296. Buffer slicing in checksum fs
Wire up TrackingByteBufferPool to vector read tests,
with test suite AbstractContractVectoredReadTest
adding test testBufferSlicing() to generate
conditions which may trigger slicing.

Only files which declare that they may
slicing buffers are permitted to return
buffers to the pool which aren't known about,
or to close the pool with outstanding entries.

So: no fix, just public declaration of behavior
and test to verify that no other streams are doing
it.