HADOOP-19611. Fix native profile build of hadoop-pipes on SLES 15.
On SLES 15, failing to add tirpc results in failures linking pthreads.
tirpc is present and appears to be the newer RPC library option (with
support for IPv6), and provides an rpc.h that Hadoop falsely thinks
comes from SunRPC. Use tirpc when available.

Tested on RedHat 8 and 9; SLES 15; and Ubuntu 20.04, 22.04, and 24.04.