HADOOP-19608. Upgrade to Junit 5.13.3
Build property trimStackTrace is set to false.

This is restores the behavior that any test failure
includes a deep stack trace rather than the first two elements.
In this version of surefire it also seems to print the full stack
of JUnit test execution and the java runtime itself.

This is only an issue for when tests fail.

Build property surefire.failIfNoSpecifiedTests is set to false.
This is required to retain the behavior where the invocation
    mvn verify -Dtest=unknown
will skip the unit tests but still run the integration tests, and a
specific property "it.test" can name the specific test to run:

  mvn verify -Dtest=unknown -Dit.test=ITestConnection