HDFS-17680. Fix redirect for https Datanode server
When the DatanodeHttpServer is configured with TLS and recieves a
request at '/', WebServlet redirects to '/index.html'. However the
infoServer does not know the scheme and creates a redirect to an 'http'
address. If only TLS is configured, this causes the redirect to fail,
and breaks links from the NameNode UI.

Adds handlers to the HTTPS Netty server to rewrite the Location header
if present to use 'https' so that the redirect works correctly.

Alternatives considered
- Add an HTTPS endpoint to infoServer. This fails because then it tries
  to decrypt the request and fails; I see no way to skip this.
- Move redirect handling into the Netty server. This would duplicate
  WebServlet logic and be more fragile as it could miss other redirects.

Unit test for DatanodeHttpServer added. Before this fix, the HTTPS
response would be 302 because URLConnection does not follow redirects
that change scheme.