HADOOP-19696. cos, huawei, aliyun cloud dependencies
* add the artifacts found with the relevant hadoop-* modules to the
  binary license
* leave all three with cloud-storage dependencies such that they
  don't include these in a pull of hadoop-cloud-storage (regression?)
* unless specific profiles cos, huawei and aliyun are declared,
  at which point they're exported by hadoop-cloud-storage and put
  into the assembly.

This avoids dealing with complex dependencies we don't want (okio, more xml
parsers,...), let making it straightforward to build a distro with it if
you want.

bundle.jar is always getting in. Do I do it here iff -Paws is set or do I
delay it until the copy to the final distro artifact tree takes place.

delay: keeps it as an export of hadoop-cloud-storage pom
early: consistent with the rest