HADOOP-19707: Add quiet-surefire profile for tests
Adds the `quiet-surefire` profile to set enableOutErrElements=false for
maven-surefire-plugin. This restores the behavior prior to Surefire 3.3
that stdout/stderr are not included in the TEST-<package>.<class>.xml
file for passing tests. The newer default behavior results in much
larger TEST-*.xml files that can be a problem for CI tools processing
them.