HADOOP-19027. S3A: S3AInputStream doesn't recover from HTTP/channel exceptions
This is a fraction of #6425 backported to the V1 SDK, which does a lot of
V2-specific translation/unwinding.

416 responses are mapped to RangeNotSatisfiableEOFException,
whose retry policy is: fail.

Classic EOFException is now retried as connection failure.
calls to read() and lazyseek all retry on this with full retry policy,
including handling of socket errors.