HADOOP-19654. third party stores and signing
The problem here is not the signing -it's that the test couldn't cope
with path-style-access as it is used the hostname of the http request
as the bucket name.

Fixed by adding a static field to note when path access is in use
and to map the name to "bucket" in such a situation.